% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AHP-package.R
\docType{package}
\name{AHPWR-package}
\alias{AHPWR}
\alias{AHPWR-package}
\title{AHPWR: Compute Analytic Hierarchy Process}
\description{
Compute a tree level hierarchy, judgment matrix, consistency index and ratio, priority vectors, hierarchic synthesis and rank. Based on the book entitled "Models, Methods, Concepts and Applications of the Analytic Hierarchy Process" by Saaty and Vargas (2012, ISBN 978-1-4614-3597-6).
}
\author{
\strong{Maintainer}: Luciane Ferreira Alcoforado \email{lucianea@id.uff.br}

Authors:
\itemize{
  \item Lyncoln Sousa \email{lyncolnsousa@id.uff.br}
  \item Orlando Celso Longo \email{orlandolongo@id.uff.br}
}

Other contributors:
\itemize{
  \item Alessandra Simão \email{alessandra_simao@id.uff.br} [contributor]
  \item Steven Dutt Ross [contributor]
  \item Leonardo Filgueira [contributor]
}

}
\keyword{internal}
