% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcula_prioridades.R
\name{calcula_prioridades}
\alias{calcula_prioridades}
\title{Calculates the priority vector of a paired matrix}
\usage{
calcula_prioridades(lista)
}
\arguments{
\item{lista}{a paired matrix list}
}
\value{
Returns a list containing priority vectors for each matrix in the read list
}
\description{
Calculates the priority vector of a paired array based on a list
}
\author{
Lyncoln Oliveira
}
