% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CI.R
\name{CI}
\alias{CI}
\title{calculates saaty's consistency index}
\usage{
CI(m)
}
\arguments{
\item{m}{is a matrice of pairwise comparison}
}
\value{
Returns  saaty's consistency index
}
\description{
Function to calculate the  saaty's consistency index
}
\examples{
x=c("c1", "c2", "c3", "c4")
y=c(3, 9, 2, 8)
m=matrix_ahp(x,y)
CI(m)

x=c("a1", "a2", "a3", "a4", "a5")
y=c(1, 9, 1.5, 8, 6)
m=matrix_ahp(x,y)
CI(m)

m=diag(16)+2-2*diag(16)
m
CI(m)
CR(m)
}
\author{
Luciane Ferreira Alcoforado
}
