% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flow_chart.R
\name{flow_chart}
\alias{flow_chart}
\title{creates a hierarchical structure}
\usage{
flow_chart(names, c, a)
}
\arguments{
\item{names}{is a vector with names for goal, criteria and choices in this exact sequence}

\item{c}{is a integer number of criteria, c>=2}

\item{a}{is a integer number of choices, a>=2}
}
\value{
Returns Diagram of hierarchies
}
\description{
Function to build the Diagram of hierarchies
}
\examples{
p=flow_chart(names=NULL, a=2, c=2)
p

p=flow_chart(names=NULL, a=2, c=3)
p+ggplot2::theme_void()

flow_chart(names=c("G", "cost", "time", "hour", "home", "beach"),c=3, a=2)

}
\author{
Luciane Ferreira Alcoforado
}
