% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formata_tabela.R
\name{formata_tabela}
\alias{formata_tabela}
\title{Format an AHP table created by the general ahp() function}
\usage{
formata_tabela(tabela, cores = "PADRAO")
}
\arguments{
\item{tabela}{AHP table created by the general ahp() function}

\item{cores}{Color pattern to format the table. If "PADRAO" returns the color pattern (green, blue, green or blue); if "GRAY" returns the default gray color; if "WHITE" returns the table without colors}
}
\value{
Returns a table formatted with background colors responsive to element priority amounts
}
\description{
Format an AHP table created by the general ahp() function
}
\author{
Lyncoln Oliveira
}
