% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formata_tabela2.R
\name{formata_tabela2}
\alias{formata_tabela2}
\title{Format an AHP table created by the general ahp() function}
\usage{
formata_tabela2(tabela, cores = "PADRAO")
}
\arguments{
\item{tabela}{AHP table created by the general ahp() function}

\item{cores}{Color pattern to format the table. If "PADRAO" returns the color pattern (green, blue, green or blue); if "GRAY" returns the default gray color; if "WHITE" returns the table without colors}
}
\value{
Retorna uma tabela formatada com cores defundo responsivas as quantidades de prioridade dos elementos
}
\description{
Format an AHP table created by the general ahp() function
}
\author{
Lyncoln Oliveira
}
