% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ler.R
\name{ler}
\alias{ler}
\title{Read an excel file containing the paired matrices and turn all your spreadsheets into a list of matrices in R}
\usage{
ler(caminho)
}
\arguments{
\item{caminho}{Address to an excel file that contains the worksheets}
}
\value{
Returns a list containing the paired arrays from the excel file
}
\description{
Function to Read an excel file containing the paired matrices and turn all your spreadsheets into a list of matrices in Re
}
\examples{
caminho  <- system.file("tests", "test_import.xlsx", package = "xlsx")
lista = ler(caminho)

}
\author{
Lyncoln Oliveira
}
