% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_ahp.R
\name{matrix_ahp}
\alias{matrix_ahp}
\title{creates a dataframe containing the judments holistic}
\usage{
matrix_ahp(x, y)
}
\arguments{
\item{x}{is a vector of names criteria or choices}

\item{y}{is a vector of weigth scale Saaty, in [1,9]}
}
\value{
Returns a judment matrice.
}
\description{
Function to build the judment matrix
}
\examples{
x=c("c1", "c2", "c3", "c4")
y=c(3, 9, 2, 8)
matrix_ahp(x,y)

x=c("a1", "a2", "a3", "a4", "a5")
y=c(1, 9, 1.5, 8, 6)
matrix_ahp(x,y)
}
\author{
Luciane Ferreira Alcoforado
}
