% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matriz_julgamento.R
\name{matriz_julgamento}
\alias{matriz_julgamento}
\title{Create paired matrix and can test saaty consistency rate}
\usage{
matriz_julgamento(n_comp, CR = TRUE, n_matrix = 1)
}
\arguments{
\item{n_comp}{Number of elements to be evaluated}

\item{CR}{If TRUE also returns the consistency rate of saaty, if FALSE returns only matrix}

\item{n_matrix}{Number of matrix to be created}
}
\value{
Returns a list with 2 positions. First position contains the paired matrices and the second position their consistency rates
}
\description{
Function that Create paired matrix and can test saaty consistency rate
}
\author{
Lyncoln Oliveira
}
