% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normaliza.R
\name{normaliza}
\alias{normaliza}
\title{generates vectors of weights for criteria and alternatives}
\usage{
normaliza(lista)
}
\arguments{
\item{lista}{is a list with judment matrices}
}
\value{
Returns  auxiliary list
}
\description{
Function that generates vectors of weights for criteria and alternatives
}
\examples{
lista = list(M1=diag(3), M2=diag(3)+4-4*diag(3))
normaliza(lista)

}
\author{
Lyncoln Oliveira
}
