% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabela_holistica.R
\name{tabela_holistica}
\alias{tabela_holistica}
\title{organizes a table with comparison matrix data}
\usage{
tabela_holistica(pesos)
}
\arguments{
\item{pesos}{vector with holistic weights for comparison or comparison matrix data}
}
\value{
Returns a table with assigned holistic weights, comparison matrix, priority vector and consistency ratio
}
\description{
Function to organizes a table with comparison matrix data
}
\examples{
pesos = c(3, 7, 9, 2)
names(pesos) = paste0("C",1:4)
tabela_holistica(pesos)

m = matrix_ahp(y=c(3,4,3,2.5), x=paste0("A",1:4))
tabela_holistica(pesos=m)

}
\author{
Luciane Ferreira Alcoforado & Orlando Longo
}
