% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transforma_tabela.R
\name{transforma_tabela}
\alias{transforma_tabela}
\title{Transforms table with numbers into percentage with 2 decimal places}
\usage{
transforma_tabela(tabela)
}
\arguments{
\item{tabela}{table to transforms}
}
\value{
Returns a transform table with percentage
}
\description{
Function to Transforms table with numbers into percentage with 2 decimal places
}
\examples{
tabela=data.frame(x=c(0.5, 0.25), y=c(0.55, 0.93))
transforma_tabela(tabela)

}
\author{
Lyncoln Oliveira
}
