% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlsx_ahp.R
\name{xlsx_ahp}
\alias{xlsx_ahp}
\title{build file with judment matrices}
\usage{
xlsx_ahp(m, file, sheet, append)
}
\arguments{
\item{m}{is a matrice of pairwise comparison}

\item{file}{is the path to the output file.}

\item{sheet}{is a character string with the sheet name.}

\item{append}{is a logical value indicating if m should be appended to an existing file. If TRUE the file is read from disk.}
}
\value{
Returns a xlsx document

#m=diag(10)
#file1 = xlsx_ahp(m, file = "Example_1.xlsx", sheet = "M1", append = FALSE)
#file2=xlsx_ahp(m, file = "Example_1.xlsx", sheet = "M2", append = TRUE)

#see file Example_1.XLSX in working directory
}
\description{
Function to buil file with judment matrices
}
\author{
Luciane Ferreira Alcoforado
}
